remove handler dependency in HttpClientTest
